Split message pump into separate module from dispatcher test
