+++ tests/dragon/test_request_dispatcher.py
@@ -25,11 +25,11 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import gc
-import io
 import logging
 import os
 import pathlib
-import socket
+import subprocess as sp
+import sys
 import time
 import typing as t
 from queue import Empty
@@ -37,7 +37,6 @@ from queue import Empty
 import numpy as np
 import pytest
 
-pytest.importorskip("torch")
 pytest.importorskip("dragon")
 
 
@@ -48,16 +47,10 @@ import torch
 
 # isort: on
 
-import dragon.channels as dch
-import dragon.infrastructure.policy as dragon_policy
-import dragon.infrastructure.process_desc as dragon_process_desc
-import dragon.native.process as dragon_process
-import torch.nn as nn
 from dragon import fli
 from dragon.data.ddict.ddict import DDict
 from dragon.managed_memory import MemoryAlloc
 
-from smartsim._core.entrypoints.service import Service
 from smartsim._core.mli.comm.channel.dragon_channel import (
     DragonCommChannel,
     create_local,
@@ -76,9 +69,7 @@ from smartsim._core.mli.infrastructure.storage.backbone_feature_store import (
 from smartsim._core.mli.infrastructure.storage.dragon_feature_store import (
     DragonFeatureStore,
 )
-from smartsim._core.mli.infrastructure.storage.feature_store import FeatureStore
 from smartsim._core.mli.infrastructure.worker.torch_worker import TorchWorker
-from smartsim._core.mli.message_handler import MessageHandler
 from smartsim.log import get_logger
 
 logger = get_logger(__name__)
@@ -89,142 +80,6 @@ try:
     mp.set_start_method("dragon")
 except Exception:
     pass
-
-
-class MiniModel(nn.Module):
-    def __init__(self):
-        super().__init__()
-
-        self._name = "mini-model"
-        self._net = torch.nn.Linear(2, 1)
-
-    def forward(self, input):
-        return self._net(input)
-
-    @property
-    def bytes(self) -> bytes:
-        """Returns the model serialized to a byte stream"""
-        buffer = io.BytesIO()
-        scripted = torch.jit.trace(self._net, self.get_batch())
-        torch.jit.save(scripted, buffer)
-        return buffer.getvalue()
-
-    @classmethod
-    def get_batch(cls) -> "torch.Tensor":
-        return torch.randn((100, 2), dtype=torch.float32)
-
-
-def load_model() -> bytes:
-    """Create a simple torch model in memory for testing"""
-    mini_model = MiniModel()
-    return mini_model.bytes
-
-
-def persist_model_file(model_path: pathlib.Path) -> pathlib.Path:
-    """Create a simple torch model and persist to disk for
-    testing purposes.
-
-    TODO: remove once unit tests are in place"""
-    # test_path = pathlib.Path(work_dir)
-    if not model_path.parent.exists():
-        model_path.parent.mkdir(parents=True, exist_ok=True)
-
-    model_path.unlink(missing_ok=True)
-
-    model = torch.nn.Linear(2, 1)
-    torch.save(model, model_path)
-
-    return model_path
-
-
-def mock_messages(
-    request_dispatcher_queue: DragonFLIChannel,
-    feature_store: FeatureStore,
-    parent_iteration: int,
-    callback_descriptor: str,
-) -> None:
-    """Mock event producer for triggering the inference pipeline"""
-    model_key = "mini-model"
-    # mock_message sends 2 messages, so we offset by 2 * (# of iterations in caller)
-    offset = 2 * parent_iteration
-
-    for iteration_number in range(2):
-        logged_iteration = offset + iteration_number
-        logger.debug(f"Sending mock message {logged_iteration}")
-
-        output_key = f"output-{iteration_number}"
-
-        feature_store[model_key] = load_model()
-
-        tensor = (
-            (iteration_number + 1) * torch.ones((1, 2), dtype=torch.float32)
-        ).numpy()
-        fsd = feature_store.descriptor
-
-        tensor_desc = MessageHandler.build_tensor_descriptor(
-            "c", "float32", list(tensor.shape)
-        )
-
-        message_tensor_output_key = MessageHandler.build_tensor_key(output_key, fsd)
-        message_model_key = MessageHandler.build_model_key(model_key, fsd)
-
-        request = MessageHandler.build_request(
-            reply_channel=callback_descriptor,
-            model=message_model_key,
-            inputs=[tensor_desc],
-            outputs=[message_tensor_output_key],
-            output_descriptors=[],
-            custom_attributes=None,
-        )
-
-        logger.info(f"Sending request {iteration_number} to request_dispatcher_queue")
-        request_bytes = MessageHandler.serialize_request(request)
-        with request_dispatcher_queue._fli.sendh(
-            timeout=None, stream_channel=request_dispatcher_queue._channel
-        ) as sendh:
-            sendh.send_bytes(request_bytes)
-            sendh.send_bytes(tensor.tobytes())
-
-        logger.info(
-            f"Retrieving {iteration_number} from callback channel: {callback_descriptor}"
-        )
-        callback_channel = DragonCommChannel.from_descriptor(callback_descriptor)
-
-        # Results will be empty. The test pulls messages off the queue before they
-        # can be serviced by a worker. Just ensure the callback channel works.
-        results = callback_channel.recv(timeout=0.1)
-        logger.debug(f"Received mock message results on callback channel: {results}")
-        time.sleep(1)
-
-
-@pytest.fixture
-def prepare_environment(test_dir: str) -> pathlib.Path:
-    """Cleanup prior outputs to run demo repeatedly"""
-    path = pathlib.Path(f"{test_dir}/workermanager.log")
-    logging.basicConfig(filename=path.absolute(), level=logging.DEBUG)
-    return path
-
-
-def service_as_dragon_proc(
-    service: Service, cpu_affinity: list[int], gpu_affinity: list[int]
-) -> dragon_process.Process:
-
-    options = dragon_process_desc.ProcessOptions(make_inf_channels=True)
-    local_policy = dragon_policy.Policy(
-        placement=dragon_policy.Policy.Placement.HOST_NAME,
-        host_name=socket.gethostname(),
-        cpu_affinity=cpu_affinity,
-        gpu_affinity=gpu_affinity,
-    )
-    return dragon_process.Process(
-        target=service.execute,
-        args=[],
-        cwd=os.getcwd(),
-        policy=local_policy,
-        options=options,
-        stderr=dragon_process.Popen.STDOUT,
-        stdout=dragon_process.Popen.STDOUT,
-    )
 
 
 def test_request_dispatcher() -> None:
@@ -279,15 +134,27 @@ def test_request_dispatcher() -> None:
 
         callback_channel = DragonCommChannel.from_local()
 
-        # create a mock client application to populate the request queue
-        msg_pump = mp.Process(
-            target=mock_messages,
-            args=(worker_queue, backbone_fs, i, callback_channel.descriptor),
+        fp = pathlib.Path(__file__).parent / "utils" / "msg_pump.py"
+        cmd = [
+            sys.executable,
+            str(fp.absolute()),
+            "--dispatch-fli-descriptor",
+            worker_queue.descriptor,
+            "--fs-descriptor",
+            backbone_fs.descriptor,
+            "--parent-iteration",
+            str(i),
+            "--callback-descriptor",
+            callback_channel.descriptor,
+        ]
+
+        popen = sp.Popen(
+            args=cmd,
+            stdout=sp.PIPE,
+            stderr=sp.PIPE,
         )
 
-        msg_pump.start()
-
-        time.sleep(1)
+        time.sleep(2)
 
         for _ in range(200):
             try:
@@ -347,8 +214,6 @@ def test_request_dispatcher() -> None:
             for mem_alloc in mem_allocs:
                 mem_alloc.free()
 
-            msg_pump.kill()
-
         request_dispatcher._active_queues[model_key].make_disposable()
         assert request_dispatcher._active_queues[model_key].can_be_removed
 
@@ -356,6 +221,8 @@ def test_request_dispatcher() -> None:
 
         assert model_key not in request_dispatcher._active_queues
         assert model_key not in request_dispatcher._queues
+
+    popen.wait()
 
     # Try to remove the dispatcher and free the memory
     del request_dispatcher
